+++ src/user/user.entity.ts
@@ -580,6 +580,16 @@ export class DbUser extends BaseEntity {
   })
   public personal_workspace_id = "";
 
+  @ApiModelProperty({
+    description: "User accepted usage terms and conditions",
+    example: false,
+  })
+  @Column({
+    type: "boolean",
+    default: false,
+  })
+  readonly accepted_usage_terms?: boolean;
+
   @ApiHideProperty()
   @OneToMany(() => DbInsightMember, (insightMember) => insightMember.member, { cascade: true })
   public insights: DbInsightMember[];